backend/requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0